kim-reporter/pyproject.toml
[project]
name = "kim-qa-reporter"
version = "0.1.0"
description = "Desktop reporter for KIM-guided couch corrections — generates clinical PDF reports from PRIME trajectory logs"
requires-python = ">=3.11"
authors = [{ name = "Kankean Kandasamy" }]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "pandas>=2.1",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "reportlab>=4.0",
    "pywebview>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pyinstaller>=6.4",
]

[project.scripts]
kim-qa-reporter = "kim_app.__main__:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["kim_app*"]

[tool.setuptools.package-data]
kim_app = ["web_dist/**/*", "pdf/assets/**/*"]
